Logic adjustments for better performance

+++ public/javascripts/server.js
@@ -7,6 +7,7 @@ const Chatmate = db.Chatmate
 const ChatRecord = db.ChatRecord
 const PublicChat = db.PublicChat
 const Subscribe = db.Subscribe
+const sequelize = db.sequelize
 const Unread = db.Unread
 const Sequelize = db.Sequelize
 const Op = Sequelize.Op
@@ -32,31 +33,48 @@ function socketConnection (io) {
         .then(userChatrooms => {
           // 將該使用者直接加入所有房間
           userChatrooms.map(item => {
-            socket.join(item)
-            socket.broadcast.to(item).emit('personal-online-notice', userId)
+            socket.join(item.id)
+            socket.broadcast.to(item.id).emit('personal-online-notice', userId)
           })
         })
       
-      // 取出該用戶所有訂閱channel
+      // 取出該用戶訂閱的所有channel
       Subscribe.findAll({
         raw: true,
-        where: { subscribing: { [Op.eq]: userId } }
+        where: { subscriber: { [Op.eq]: userId } }
       })
       .then(subscribeChannel => {
         // 直接加入訂閱channel
         subscribeChannel.map(item => {
-          const roomName = 's' + item.id
+          const roomName = 's' + item.subscribing
           socket.join(roomName)
         })
       })
 
-      // 取出user所有未讀取的通知
+      // 取出user所有未讀取的通知數量
       Unread.findAll({
+        raw: true,
         where: { receiveId: { [Op.eq]: userId } },
-        attributes: [sequelize.fn('COUNT', sequelize.col('id')), 'unreadCount']
+        attributes: [[sequelize.fn('COUNT', sequelize.col('id')), 'unreadCount']]
       })
       .then(unreads => {
         socket.emit('notices', unreads)
+      })
+
+      socket.on('new-tweet', async (userId) => {
+        // 取出訂閱該使用者的清單
+        const subscribers = await Subscribe.findAll({
+          raw: true,
+          where: { subscribing: { [Op.eq]: userId } },
+          attributes: ['subscriber']
+        })
+        
+        // 對訂閱者發送通知(有新推文時)
+        subscribers.map(element => {
+          const roomId = 's' + element.subscriber
+          socket.join(roomId)
+          socket.broadcast.to(roomId).emit('notices', 1)
+        });
       })
 
       // 當user觸發讀取通知
@@ -69,12 +87,20 @@ function socketConnection (io) {
 
           socket.emit('read-notice', unreads)
 
-          await Unread.destroy({
-            where: { receiveId: { [Op.eq]: userId } }
-          })
+          if (unreads.length) {
+            await Unread.destroy({
+              where: { receiveId: { [Op.eq]: userId } }
+            })
+          }
         }
         catch (err) {
           console.log(err)
+        }
+      })
+
+      socket.on('notices', ({ targetId }) => {
+        if (onlineList[targetId]) {
+          socket.to(onlineList[targetId]).emit('notices', 1)
         }
       })
 
@@ -94,7 +120,7 @@ function socketConnection (io) {
           })
 
           const publicChatRecord = await PublicChat.findAll({
-            include: { model: User, as: 'speaker', attributes: ['name', 'account', 'avatar'] }
+            include: { model: User, as: 'user', attributes: ['name', 'account', 'avatar'] }
           })
 
           // 公告使用者上線
@@ -115,16 +141,11 @@ function socketConnection (io) {
       })
 
       // 監聽公開聊天室訊息並廣播
-      socket.on('public-msg', async ({ userId, message }) => {
-        try {
+      socket.on('public-msg', ({ userId, message }) => {
           PublicChat.create({
             speakerId: userId,
             chatContent: message
           })
-        }
-        catch (err) {
-          console.log(err)
-        }
 
         socket.broadcast
           .to('public-room')
